feat(Args): Parametric the function by adding bems
